checkpoint before combine the sign up page
