+++ src/main/resources/data.sql
@@ -1,3 +1,3 @@
-INSERT INTO article(id, title, content) VALUES(1, 'aaaa', '1111');
-INSERT INTO article(id, title, content) VALUES(2, 'bbbb', '2222');
-INSERT INTO article(id, title, content) VALUES(3, 'cccc', '3333');
+INSERT INTO article(id, login_ID, login_pw) VALUES(1, 'aaaa', '1111');
+INSERT INTO article(id, login_ID, login_pw) VALUES(2, 'bbbb', '2222');
+INSERT INTO article(id, login_ID, login_pw) VALUES(3, 'cccc', '3333');
