+++ src/main/resources/templates/articles/login.mustache
@@ -21,13 +21,13 @@
 							<legend class="pure-steps_group-step_legend">Sign in</legend>
 							<div>
 							<p class="pure-steps_group-step_item flexy-item flexy-column reverse">
-								<input type="text" placeholder="Type your ID here" value="" name="ID">
+								<input type="text" placeholder="Type your ID here" value="" name="login_ID">
 								<label class="form-label">ID</label>
 							</p>
 							</div>
 							<div>
 							<p class="pure-steps_group-step_item flexy-item flexy-column reverse">
-								<input type="password" placeholder="Type your password here" value="" name="password">
+								<input type="password" placeholder="Type your password here" value="" name="passWord">
 								<label class="form-control">Password</label>
 							</p>
 							</div>
