+++ src/main/resources/templates/articles/new.mustache
@@ -3,11 +3,11 @@
 <form class="container" action="/articles/create" method="post">
 	<div class="mb-3">
 	  <label class="form-label">제목</label>
-	  <input type="text" class="form-control" name="title">
+	  <input type="text" class="form-control" name="login_ID">
 	</div>
 	<div class="mb-3">
 	  <label class="form-label">내용</label>
-	  <textarea class="form-control" rows="3" name="content"></textarea>
+	  <textarea class="form-control" rows="3" name="passWord"></textarea>
 	</div>
 	<button type="submit" class="btn btn-primary">Submit</button>
 	<a href="/articles">Back</a>
